requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-ninja==1.1.0
